🎫 transform class to function level component
