+++ src/components/info/info.jsx
@@ -124,3 +124,4 @@ export const Info = (props) => {
 
 
 
+export default Info;
